refactor: move store type
