+++ src/lib/features/instance-stats/instance-stats-service.ts
@@ -8,7 +8,7 @@ import {
 } from '../../types/stores';
 import { IContextFieldStore } from '../../types/stores/context-field-store';
 import { IEnvironmentStore } from '../../types/stores/environment-store';
-import { IFeatureToggleStore } from '../../types/stores/feature-toggle-store';
+import { IFeatureToggleStore } from '../feature-toggle/types/feature-toggle-store-type';
 import { IGroupStore } from '../../types/stores/group-store';
 import { IProjectStore } from '../../types/stores/project-store';
 import { IStrategyStore } from '../../types/stores/strategy-store';
