+++ src/lib/middleware/rbac-middleware.test.ts
@@ -4,7 +4,7 @@ import * as perms from '../types/permissions';
 import { IUnleashConfig } from '../types/option';
 import { createTestConfig } from '../../test/config/test-config';
 import ApiUser from '../types/api-user';
-import { IFeatureToggleStore } from '../types/stores/feature-toggle-store';
+import { IFeatureToggleStore } from '../features/feature-toggle/types/feature-toggle-store-type';
 import FakeFeatureToggleStore from '../../test/fixtures/fake-feature-toggle-store';
 import { ApiTokenType } from '../types/models/api-token';
