+++ src/lib/types/stores.ts
@@ -4,7 +4,7 @@ import { IFeatureTypeStore } from './stores/feature-type-store';
 import { IStrategyStore } from './stores/strategy-store';
 import { IClientApplicationsStore } from './stores/client-applications-store';
 import { IClientInstanceStore } from './stores/client-instance-store';
-import { IFeatureToggleStore } from './stores/feature-toggle-store';
+import { IFeatureToggleStore } from '../features/feature-toggle/types/feature-toggle-store-type';
 import { IContextFieldStore } from './stores/context-field-store';
 import { ISettingStore } from './stores/settings-store';
 import { ISessionStore } from './stores/session-store';
